데이터 타입 통일, url 통일, partner image 생성

+++ festival/src/main/java/com/example/festival/comment/dto/CommentDto.java
@@ -13,10 +13,10 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CommentDto {
-    private Long commentId;
+    private Integer commentId;
     private String nickname; //댓글 작성자
     private String address; //작성자 주소
-    private Long partnerId; //게시글
+    private Integer partnerId; //게시글
     private String content; //내용
     private Timestamp createdAt; //작성일
     private List<ReplyDto> replyDtos; //댓글에 달린 답글
